requirements.txt
flask
pandas
python-dotenv
sentence-transformers
pinecone